Fix desugaring when LOCAL_SDK_VERSION is not set
When LOCAL_SDK_VERSION is not set, frameworks.jar is put
in the classpath for javac.  If it is also in the classpath
(as opposed to the bootclasspath) for desugar then desugar
tries to rewrite classes that implement interfaces with
default methods to contain calls to the generated companion
classes for those interfaces.  frameworks.jar will not
contain the companion classes, which causes proguard to
fail.

Use --bootclasspath_entry for the TARGET_DEFAULT_JAVA_LIBRARIES
so that desugar knows those libraries will not be desugared,
and use --classpath_entry for the rest.  Also remove the static
libraries from the classpath, they have already been merged into
the input jar.

Test: m checkbuild
Change-Id: If30943efcaea44b2db1b38a6c1d558113324757f