javac 9 -target 1.9: Move bootclasspath jars onto classpath.
When building with OpenJDK 9's javac with -source 1.9 -target 1.9,
a runtime image (with system modules) is used via the --system
command line flag, instead of --bootclasspath.
The runtime image only contains the default libraries (libcore)
but is missing other libraries that might be on the bootclasspath
for a particular build target.

This CL fixes compilation for this case by adding the missing
jars onto the classpath instead.

This already used to work but was broken by CL [1]. I attempted
to conditionally revert some of that CL's changes to java_common.mk
for the case of language level 1.9, but couldn't get it to work.
Since Android targets outside libcore currently all use the same
system modules (core-system-modules), and because there doesn't
seem to be any issue with classes from the system modules also
appearing on the classpath, this CL chooses this simpler approach
instead.

[1] http://r.android.com/519552

Bug: 38177569
Test: Treehugger

Change-Id: I48fd11aac9b310bfa58dee0f9cfb3ef33f10bca8