Basic support for device or product specific modules
ODM partition is the device-specific extension of the vendor partition
and OEM partition is the product-specific extension of the system
partition. For example, an HAL for an off-chip peripheral which isn't
implemented by an SoC vendor but by an ODM, can go into the ODM
partition. Likewise, an Android app that is only for a specific network
carrier can go into the OEM partition instead of the system partition.

As the first step, the path to the odm and oem partitions are exported
from make to the soong world. The partition paths are /odm and /oem when
the partitions exist, and are /vendor/odm and /system/oem when they
don't exist. Like /system/vendor is a symlink to /vendor, two symlinks
/vendor/odm and /system/oem are created when the partitions exist.

Bug: 68187740
Test: WIP

Change-Id: I8957551dd24d21e94db70a85e9d9c9aa6ac1e0c7