*.cfi.vendor is recognized as native:vndk
The cfi variants have not been recognized as vndk because of the .cfi
suffix. This becomes a problem when a sanitized VNDK lib (ex:
libstagefright_omx.so) is using a private VNDK lib (ex: libgui.so)
because the former is falsely identified as native:vendor which is not
allowed to link to the private VNDK libs.

Bug: 38298593
Test: m -j libstagefright_omx on walleye

Change-Id: I0aefbba4a05373a3dc987c70b9db14345bd3e418