Update patch set 1
Patch Set 1:

Since if the build fails , that rule will not be guaranteed to get triggered (and that is when we really want the files copied over to dist). This script can be called from the build_cmd on build servers to copy the abidiffs from the out dir to the dist dir in the case of a build failure.

Patch-set: 1
Reviewer: Gerrit User 1121338 <1121338@85c56323-6fa9-3386-8a01-6480fb634889>