Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> frameworks/opt/net/wifi/tests/wifitests/Android.mk uses inner classes and so they intentionally specify a filter of "com.android.server.wifi.Test*" to match either "com.android.server.wifi.Test" or "com.android.server.wifi.Test$*"
> 
> Maybe worth supportting "*" at the end even when not following a "."

Unfortunate, but OK.  The semantics will end up a little weird.  'a.b.*' will match 'a.b.c.d' (because it will translate to a/b/**/*.class), but 'a.b.c*' will not (it will need to translate to a/b/c*.class).

Patch-set: 2
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>