Fall back to normal split if imgdiff fails when splitting large apks
Imgdiff expects the input files to be valid zip archives. During the
split of large apks, imgdiff may fail when there's a hole in the input
file potentially due to the blocks allocation of mke2fs. This CL solves
the issue by falling back to normal split in case of a imgdiff split
failure. The split transfer will then use bsdiff instead.

Bug: 69624507
Test: generate the incremental package for sawfish and check the
transfers.
Change-Id: I4882452378123e60ad3434053b57f33e53ac4b82