Update patch set 7
Patch Set 7: Presubmit-Verified+1

=== Finished presubmit run: L97300000123455479 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/540425?revision=7




TreeHugger finished with: 19 passed.







WARNING: The tip-of-tree(4473390) for git_master/errorprone_fastbuild3e_linux may be broken.
The presubmit build passes with the last-known-good-build(4472965), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4473390) for git_master/gce_x86_phone-userdebug_fastbuild3c_linux may be broken.
The presubmit build passes with the last-known-good-build(4473011), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.









WARNING: The tip-of-tree(4473390) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4473011), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4473390) for git_master/sailfish-userdebug_fastbuild3_linux may be broken.
The presubmit build passes with the last-known-good-build(4473011), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4473390) for git_master/test_suites_x86_64_fastbuild3d_linux may be broken.
The presubmit build passes with the last-known-good-build(4473011), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.






Please accept my sincere yet pre-programmed message of undying praise.

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 7
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=+1