Update patch set 6
Patch Set 6: Presubmit-Verified+1

=== Finished presubmit run: L64200000123479719 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/547360?revision=6




TreeHugger finished with: 23 passed.







WARNING: The tip-of-tree(4473590) for git_master/errorprone_fastbuild3e_linux may be broken.
The presubmit build passes with the last-known-good-build(4472965), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4473590) for git_master/gce_x86_phone-userdebug_fastbuild3c_linux may be broken.
The presubmit build passes with the last-known-good-build(4473011), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.









WARNING: The tip-of-tree(4473590) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4473011), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.






WARNING: The tip-of-tree(4473590) for git_master/sailfish-userdebug_fastbuild3_linux may be broken.
The presubmit build passes with the last-known-good-build(4473011), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4473590) for git_master/taimen-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4473011), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4473590) for git_master/test_suites_x86_64_fastbuild3d_linux may be broken.
The presubmit build passes with the last-known-good-build(4473011), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.






I am TreeHugger, and I have hugged this change!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 6
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=+1