Update patch set 1
Patch Set 1:

(1 comment)

Also added a note on b/62019611. But that's besides the point. A vendor can delete what's in the conditional, the check needs to be based on shipping api level and other standard measures in the OTA server. If that's not done, you can never be sure anyway.

Patch-set: 1
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>