Moving recovery resources from /system to /vendor
For non-A/B devices:
  boot.img + recovery-from-boot.p + recovery-resource.dat = recovery.img

The generated recovery.img can be used to update the /recovery partition.
With system-only OTA introduced in project Treble, we should move
related resources to vendor partition. So they won't be affected by
system-only OTA.

Specifically, this CL moves the following files from /system to /vendor:
  - /system/etc/recovery-resource.dat -> /vendor/etc/recovery-resource.dat
  - /system/recovery-from-boot.p -> /vendor/recovery-from-boot.p

Bug: 68319577
Test: boot a device
Change-Id: I2a848dbb711576e0789419c81c2ac58aabaa20fa