Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The logic seems wrong to me; in my understanding, this CL tries to avoid a "free" command in the middle of two "stash" commands with the same SHA.
> Example:
> 1. stash 0x100 into A;
> 2. use A for some operations;
> 3. free A;      ---> (avoid this cmd)
> ...
> 4. stash 0x7100 into A;
> 5. use A for some operations;
> 6. free A;
> 
> But in the current implementation, we skip stashing the blocks if a stash with the same SHA exists, even if the block contents are different (i.e. a hash collision). As a result, we'll use the stash 0x100 during the update while we actually expect 0x7100.
> And we should keep the current behavior about stash overwriting, since the source blocks to stash may no longer be intact for interrupted updates.

The hash collision doesn't look like a major concern to me, because we have been relying on the SHA to identify a "unique" stash. Here's the code from ReviseStashSize():

    for xf in self.transfers:
      ...
      # xf.stash_before generates explicit stash commands.
      for stash_raw_id, sr in xf.stash_before:
        # Check the post-command stashed_blocks.
        stashed_blocks_after = stashed_blocks
        sh = self.src.RangeSha1(sr)
        if sh not in stashes:
          stashed_blocks_after += sr.size()

        if stashed_blocks_after > max_allowed:
          ...
        else:
          # Update the stashes map.
          if sh in stashes:
            stashes[sh] += 1
          else:
            stashes[sh] = 1
          stashed_blocks = stashed_blocks_after

We just consider later stashed blocks the same as an existing one, based on the SHA. So it's possible to have other blocks (e.g. 0x200, 0x300) stashed to the same slot as 0x100 and 0x7100. I guess from cryptography algorithms POV, SHA-1 hash collision is not a concern in BBOTA.

However, there're other issues regarding this change.

The added ReviseStashCount() changes the lifecycle of stashes, which invalidates the max stash size computation in ReviseStashSize(). Because now we're keeping some stashes longer, the occupied space could exceed the free space limit and lead to update failure. This ReviseStashCount() can't come as a fix-up after ReviseStashSize().

After fixing the above issue, this change could also affect the package size for the same reason. Because if it runs out of cache space after removing those "free" commands, some later commands need to be converted into "new" commands, at the cost of increasing package size. This is the limit of this approach.

Patch-set: 1
Reviewer: Gerrit User 1056365 <1056365@85c56323-6fa9-3386-8a01-6480fb634889>