Add java:core as a new link type
java:core is for a java module that isn't using framework libraries but
only core java libraries like core-oj and core-libart.

A module of this type can be linked from any other link type, but this
type cannot link to other types.

This new type is created in the course of making all link-type check
warnings as errors. We currently have many violations of java:sdk ->
java:platform. A module has usually been identified as java:platform not
because it is using framework libraries, but because LOCAL_SDK_VERSION
isn't set. We can workaround the violation by setting LOCAL_SDK_VERSION
to some value but that isn't theoretically correct and in some cases, it
isn't totally undesirable. For example, some test support libraries in
libcore shouldn't be built with API stubs as they need to be built
without framework source code. So, this new java:core is to model such
libraries which intentionally do not have dependencies to the SDK or
framework libraries.

Test: m -j checkbuild shows less link-type check warnings
Change-Id: Ibc5caabcfc23ba82468f104ec3290a8a0cd04c92