Update patch set 6
Patch Set 6: Presubmit-Verified+1

=== Finished presubmit run: L66000000135727282 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/583852?revision=6&workplanId=L66000000135727282




TreeHugger finished with: 18 passed.







WARNING: The tip-of-tree(4552653) for git_master/gce_x86_phone-userdebug_fastbuild3c_linux may be broken.
The presubmit build passes with the last-known-good-build(4552429), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.








WARNING: The tip-of-tree(4552653) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4552440), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4552653) for git_master/sailfish-userdebug_fastbuild3_linux may be broken.
The presubmit build passes with the last-known-good-build(4552429), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4552653) for git_master/taimen-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4552340), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4552653) for git_master/walleye-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4552378), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



Congratulations!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 6
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=+1