Support LOCAL_SDK_VERSION := core_current
core_current is a pseudo SDK version which is a core Java API subset of
the Android API. It is expected to be mainly used for external Java
projects which are agnostic to Android; such as junit, guava, etc.

A module built with this SDK version can only link to java modules of
the same kine. It can't depend on modules built with LOCAL_SDK_VERSION
:= current or without LOCAL_SDK_VERSION. (For now, violating this only
generates warnings, but this will be errors in the future)

Bug: 72206056
Test: WIP
Change-Id: I34a9696393aa6704fd6684a40ea5b05d3fb46b23