Fix recovery upgrade on AVB enabled non-A/B devices.
This is to update the SHA values in install-recovery.sh etc in /system
that are needed for recovery upgrade on non-A/B.
With AVB enabled, boot and recovery images are modified due to avb
footer addition after the SHAs are stored, causing recovery upgrade
failure. Do an additional MakeRecoveryPatch call to fix it.

Test: validate_target_files.py <target_files.zip> fails till now on
      AVB enabled non-A/B devices.
      It PASSes with this change.

Change-Id: I8d736d5b309f8a3f9615ef48663c55ef4239a429