Uncompress dex of jars in vendor image
The goal is to make it unnecessary to whitelist jars used by priv apps
in PRODUCT_LOADED_BY_PRIVILEGED_MODULES.

This change covers both java libraries installed to /vendor, and
prebuilt jars that are copied to both /system.  For java libraries that
are installed to /system, they are currently pre-dexopted and dex is
strpped.  Effectivly, we should have no compressed dex anywhere.

Test: marlin and walleye built, vendor/framework/*.jar contain
      uncompressed dex
Test: Check that size of jars in vendor/framework/ are in the order of
      hundreds K on marlin, and even smaller on walleye
Bug: 63920015
Bug: 30972906

Change-Id: Ib5e9af350da16e02c2e104478f8432449b5f7455