Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Is there always going to be a single whitelist (CTS verified?), or could this be vendor/device specific? If it could change, it would make more sense to use PRODUCT_COPY_FILES so that it can be overridden without making source changes to frameworks/base.
> 
> Not clear yet, but:
> - The change to read the whitelist supports reading from various locations/partitions anrd merging at runtime: https://android-review.googlesource.com/c/platform/frameworks/base/+/620447/3/core/java/com/android/server/SystemConfig.java

From that, it sounds like the current approach would be reasonable, since you could install per-application xml files to add more entries to this list.

> - The longer term plan is to construct the whitelist at build time, based on information about the packages going into the build & what API(s) they build against

This is not supported by the build system, as we cannot get a list of packages that are going to be installed for a particular build. The naive approach is easily foiled by strange dependencies and user requests. We hope to make this explicit in the future, but that's likely going to be a while.

What would be feasible would be for each of the packages (that meet a set of requirements) to depend on installing a per-app xml file. That way the xml file would be installed if the app did end up getting installed, but we wouldn't be adding information to the image if the app wasn't installed.

> So when we're done, there should be no checked in whitelist at all.

Patch-set: 1
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>