Update patch set 2
Patch Set 2:

This won't necessarily cause the protoc command to be re-run if the options file is added, and definitely won't cause it to rerun if the options file is removed. It's rather tough to handle optional reads properly using Make/Ninja rules.

The most consistent option would be for us to read the dependency files written by aprotoc, but they're not current in a format understood by ninja, and they don't contain the option file :-(  I'll look to see if we can post-process that file.

Patch-set: 2
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>