releasetools: Add PropertyFiles for non-A/B OTA packages.
This CL exposes ota-property-files flag for non-A/B OTA packages.
Currently the line only contains the info for the METADATA entry, for
example "ota-property-files=metadata:69:286". This allows system updater
to just download the METADATA entry based, as opposed to downloading the
entire package, to learn about the info regarding the OTA package. Note
that this requires the OTA server side support to pass down the flag
along with the update URL.

Bug: 74210298
Test: WIP
Change-Id: I1482c587e18ea7101c8328777ea988c2d8ca06ac