Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> End goal is to remove these from all variants of GSI. Do we have a bug tracking that?

As you commented, we have a bug for qualcomm (b/64905218). As the directories will be moved to vendor partition in P, we don't need to merge this CL in P but in oc-mr1-treble-dev.

I will abandon this CL and upload another to oc-mr1-treble-dev for b/73720182.

Patch-set: 1
Reviewer: Gerrit User 1146974 <1146974@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0